SOURCEWEB-1 Add include path for gcc-4.8 #1
gcc-4.8 has some weird include path problems. Although additional
investigation as to why the projects have so much trouble including
the path should be investigated further a temporary workaround is to
directly include the necessary include path to find <bits/config.h>.

This version of the code is so that it's easier to cherry-pick this
change onto another repo.